Use integer constants for dat sp33d
This is stupid, but conversion of literals like 123 in code to an
integer value happens at runtime and can be very slow. Doing it for hot
paths means real perf impact so by separating some integer values out
into constants, we can speed up our code.

I think this yields like a 30% boost on my machine, but I haven't
benchmarked that carefully. Been trying lots of things, so could be
15%.. but it's quite certainly not less. The differene is significant.