Navigate to right schema node
We talked about having one TsdbWriter per device and that we could place
the cursor at the right place in the schema and the TsdbWriter could
take it from there. That would work for config or live-status (oper)
data but not for RPCs as there is no common root for all RPCs on a
device. The /devices/device/rpc node is in NSO namespace, not the NED /
device. This quirk is only an issue on JUNOS where we do RPCs, but... it
is an issue nonetheless.

For RPCs, the schema cursor needs to be placed differently based on
which RPC was queried. The RPC reply does not indicate which RPC was
called, so we have to keep track of that separately too, thus passing a
cursor per RPC reply.

Navigating the cursor is a huge hack right now as there's some issue
with the JUNOS NED namespace not being properly loaded into the
NamespaceMaps so we cannot do proper looks of ITag to HTag. I worked
around it, see dumb_push() function.