Workaround for problem with generic in acton tip
A recent change in acton, 223a8dd152c26c25d85a577ab79a4bb8ddfdf420 to
generate wrappers for hidden functions, somehow causes a breakage on
this generic function. We should fix the compiler but in the mean time
we work around it by changing the type here to the only type it is
currently used for.