Add NSO skeleton
This includes a NSO in Docker skeleton so that we can easily build our
own NSO images for the purpose of testing.

I've included a JUNOS NED, currently with just the YANG model for the
interface related RPCs. The YANG model comes from Junipers github repo
with YANG models.

I used the ned skeleton but modified it somewhat to produce an image
called nso rather than testnso and to tag it...