Do RPC to config false mapping for get-interface-information
When we write to the TSDB we write config false data and the YANG schema
describes that structure. We want to be able to give the YANG model to a
consumer so they in turn can query the TSDB.

Since for JUNOS we do RPC polling but it doesn't make any sense and
there is no natural representation of RPCs in a TSDB, we do a conceptual
translation from RPC output to a config false tree. Rather than just
doing that in code, I've applied the translation mapping for the
get-interface-information RPC by creating it as a config false node in
the schema. This means we have a materialization of it that's easy to
understand.

We should eventually build a tool to translate RPC models for us.