Simplify core dump debugging
Run test as own user so the resulting core file is owned by ourself
outside of the container. This means we have to create a user for
ourselves inside the container since it normally only comes with root
and some system accounts.

Also, change current directory to the bin path, so the core file ends up
next to the executable binary.