Add prsrc() to print source for XNode tree
This allows us to print what the source code looks like for a whole data
structure. It's similar to __str__ or __repr__ output, but here we have
correct quoting so it's actually valid source code, and do indent +
newlines so it looks OK.

Print some data structure and copy it into your test cases!