Add startup delay to get-poller
On my machine, the get-poller or rather the SSH client process, dies
with exit code 7 if there is no delay here. I have no idea why but
adding a small delay fixes it. Since it's only on startup and this
significantly simplifies developing, I'm putting this in for now but it
should totally go away.