Move NED to get it exported
In the MPLS top NSO, we want to use this minified NED that only supports
the IP SLA model on IOS XE, so moving it from test-packages to packages.

This is an interim kind of thing, it'll likely change again in the
future and then we can move this package back.