Avoid extra log handler in PRW client
We don't want an extra log handler actor, just a local logger that is
using the fed in handler. The purpose of a handler is to be able to
collect the logs from multiple actors under a single control (and it's
also reflected in the path). With a log handler per Client, we don't
really get that, it's just an extra jump for not much gain.