Adapt to optional_str removal in Acton
The str class now takes an ?value so calling str() on an optional type now
works. Previously it did not which led to the creation of helper functions like
optional_str that would solve this. Since this change, such helper functions
are now unnecessary and we should remove them and use the idiomatic str()
instead.

Similarly for optional_repr and logging functions now also allow optional
values so we can simplify quite a lot of code.

Use more idiomatic Acton string interpolation as well.